Crate symbolic_polynomials

This is a module for manipulation of symbolic polynomials overs the integers. Unfortunately, at this stage since Rust does not support integers as Generic types, the only possible implementation of SymMonomials is with a vector. This in terms means that it can not implement the Copy trait, for which reason currently all operations are defined for references to the actual structures. If this changes in the future I will update the module accordingly.

Any examples of usage can be found under the relevant methods in each structure documentation.

Structs

SymMonomial
An implementation for a monomial over the integers.

SymPolynomial
A symbolic polynomial over the integers.

Constants

DEFAULT_VARIABLES
The default number of symbolic variables for now is set 10