Enum symbolic_polynomials::Composite

pub enum Composite<I, C, P> where
    I: Id,
    C: Coefficient,
    P: Power,  {
    Variable(I),
    Floor(Rc<Polynomial<I, C, P>>, Rc<Polynomial<I, C, P>>),
    Ceil(Rc<Polynomial<I, C, P>>, Rc<Polynomial<I, C, P>>),
    Min(Rc<Polynomial<I, C, P>>, Rc<Polynomial<I, C, P>>),
    Max(Rc<Polynomial<I, C, P>>, Rc<Polynomial<I, C, P>>),
}

A composite expression (tagged union) of a variable or an irreducible function (floor, ceil, max, min).

Variants
Variable(I)Floor(Rc<Polynomial<I, C, P>>, Rc<Polynomial<I, C, P>>)Ceil(Rc<Polynomial<I, C, P>>, Rc<Polynomial<I, C, P>>)Min(Rc<Polynomial<I, C, P>>, Rc<Polynomial<I, C, P>>)Max(Rc<Polynomial<I, C, P>>, Rc<Polynomial<I, C, P>>)
Methods

impl<I, C, P> Composite<I, C, P> where
I: Id,
C: Coefficient,
P: Power,

fn eval(&self, values: &HashMap<I, C>) -> Result<C, (I, String)>

Evaluates the Composite given the provided mapping of identifiers to value assignments.

fn to_code<F>(&self, format: &F) -> String where
F: Fn(I) -> String,

Returns a code equivalent string representation of the Composite. The format specifies a function how to render the identifiers.

fn unique_identifiers(&self, unique: &mut HashSet<I>)

Fills into the HashSet all of the identifiers used in this Composite.