Struct symbolic_polynomials::Monomial

pub struct Monomial<I, C, P> where
    I: Id,
    C: Coefficient,
    P: Power,  {
    pub coefficient: C,
    pub powers: Vec<(Composite<I, C, P>, P)>,
}

A symbolic monomial represented as C * a_1^p_1 * a_2^p_2 * ... * a_n^p_n.

Fields
coefficient: C

The constant coefficient C

powers: Vec<(Composite<I, C, P>, P)>

A vector of the pairs (a_i, p_i), where a_i is a Composite expression.

Methods

impl<I, C, P> Monomial<I, C, P> where
I: Id,
C: Coefficient,
P: Power,

fn up_to_coefficient<T: AsRef<Monomial<I, C, P>>>(&self, other: T) -> bool

Returns true if the the two monomials are equal, ignoring their coefficients.

fn is_constant(&self) -> bool

Returns true only if the monomial does not depend on any symbolic variables.

fn eval(&self, values: &HashMap<I, C>) -> Result<C, (I, String)>

Evaluates the Monomial given the provided mapping of identifiers to value assignments.

fn to_code<F>(&self, format: &F) -> String where
F: Fn(I) -> String,

Returns a code equivalent string representation of the Monomial. The format specifies a function how to render the identifiers.

fn unique_identifiers(&self, unique: &mut HashSet<I>)

Fills into the HashSet all of the identifiers used in this Monomial.

impl<I, C, P> Monomial<I, C, P> where
I: Id,
C: Coefficient,
P: Power,

fn checked_div(&self, rhs: &Monomial<I, C, P>) -> Option<Self>

If the the monomial is divisible by rhs than returns the result of that division, otherwise None.