Struct symbolic_polynomials::Polynomial

pub struct Polynomial<I, C, P> where
    I: Id,
    C: Coefficient,
    P: Power,  {
    pub monomials: Vec<Monomial<I, C, P>>,
}

A symbolic polynomial represented as m_1 + m_2 + ... + m_n.

Fields
monomials: Vec<Monomial<I, C, P>>

A vector of the monomials m_i.

Methods

impl<I, C, P> Polynomial<I, C, P> where
I: Id,
C: Coefficient,
P: Power,

fn is_constant(&self) -> bool

True only if the polynomial is constant and does not depend on any symbolic variables.

fn eval(&self, values: &HashMap<I, C>) -> Result<C, (I, String)>

Evaluates the Polynomial given the provided mapping of identifiers to value assignments.

fn to_code<F>(&self, format: &F) -> String where
F: Fn(I) -> String,

Returns a code equivalent string representation of the Polynomial. The format specifies a function how to render the identifiers.

fn div_rem(
&self,
rhs: &Polynomial<I, C, P>
) -> (Polynomial<I, C, P>, Polynomial<I, C, P>)

Returns the result of the polynomial division with rhs as well as the reminder. Note that this division depends on the ordering of the variable variables type I as explained in Wikipedia.

fn checked_div(&self, rhs: &Polynomial<I, C, P>) -> Option<Polynomial<I, C, P>>

If the the polynomial is divisible by rhs than returns the result of that division, otherwise None.

fn unique_identifiers(&self, unique: &mut HashSet<I>)

Fills into the HashSet all of the identifiers used in this Polynomial.